// backend/src/integrations/apollo/apollo.generated.ts
export interface operations {
    "people-enrichment": {
        parameters: {
            query?: never
            header?: never
            path?: never
            cookie?: never
        }
        requestBody?: never
        responses: {
            /** @description Success */
            200: {
                headers: {
                    [name: string]: unknown
                }
                content: {
                    "application/json": {
                        person?: {
                            id?: string
                            first_name?: string
                            last_name?: string
                            name?: string
                            title?: string
                            seniority?: string
                            email?: string
                            email_status?: string
                            linkedin_url?: string
                            city?: string
                            state?: string
                            country?: string
                            organization?: {
                                id?: string
                                name?: string
                                website_url?: string
                                primary_domain?: string
                                industry?: string
                                estimated_num_employees?: number
                            } & {
                                [key: string]: unknown
                            }
                        } & {
                            [key: string]: unknown
                        }
// ...
                    }
                }
            }
        }
    }
    "bulk-people-enrichment": {
        parameters: {
            query?: never
            header?: never
            path?: never
            cookie?: never
        }
        requestBody?: never
        responses: {
            /** @description Success */
            200: {
                headers: {
                    [name: string]: unknown
                }
                content: {
                    "application/json": {
                        matches?: ({
                            id?: string
                            first_name?: string
                            last_name?: string
                            name?: string
                            title?: string
                            seniority?: string
                            email?: string
                            email_status?: string
                            linkedin_url?: string
                            city?: string
                            state?: string
                            country?: string
                            organization?: {
                                id?: string
                                name?: string
                                website_url?: string
                                primary_domain?: string
                                industry?: string
                                estimated_num_employees?: number
                            } & {
                                [key: string]: unknown
                            }
                        } & {
                            [key: string]: unknown
                        })[]
// ...
                    }
                }
            }
        }
    }
// ...
    "people-api-search": {
        parameters: {
            query?: never
            header?: never
            path?: never
            cookie?: never
        }
        requestBody?: never
        responses: {
            /** @description Success */
            200: {
                headers: {
                    [name: string]: unknown
                }
                content: {
                    "application/json": {
                        people?: ({
                            id?: string
                            first_name?: string
                            last_name_obfuscated?: string
                            title?: string | null
                            has_email?: boolean
                            organization?: {
                                name?: string
                            } & {
                                [key: string]: unknown
                            }
                        } & {
                            [key: string]: unknown
                        })[]
// ...
                    }
                }
            }
        }
    }
    "organization-jobs-postings": {
        parameters: {
            query?: never
            header?: never
            path?: never
            cookie?: never
        }
        requestBody?: never
        responses: {
            /** @description Success */
            200: {
                headers: {
                    [name: string]: unknown
                }
                content: {
                    "application/json": {
                        organization_job_postings?: ({
                            id?: string
                            title?: string
                            url?: string
                            city?: unknown
                            state?: unknown
                            country?: string
                            posted_at?: string
                            last_seen_at?: string
                        } & {
                            [key: string]: unknown
                        })[]
// ...
                    }
                }
            }
        }
    }
}

export type OperationId = keyof operations

type JsonContent<T> = T extends { content: infer Content } ? (Content extends { "application/json": infer Body } ? Body : never) : never

type Responses<Id extends OperationId> = operations[Id]["responses"]

export type SuccessResponseBody<Id extends OperationId> = {
    [Status in keyof Responses<Id>]: `${Status & (string | number)}` extends `2${string}` ? JsonContent<Responses<Id>[Status]> : never
}[keyof Responses<Id>]